Fix failure to log in when requesting binary files
The problem was that the authenticated fetch for establishing
a session cookie was not fully completed when reloading the page.
This caused a new (non-authenticated) session cookie to be created.

This fix changes this fetch request for the session cookie to be a
a header request, so that the request is always finalized
before the reload is called.
An alternative (but less efficient) fix would be to do a regular
request, and consume the whole body before reloading the page.

Closes #878